점수 표시 관련, 사양서대로 수정.
