+++ Assets/CoreAssets/Animators.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7be9a72f667eb1145b6328dddff84b99
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
