+++ Assets/CoreAssets/Animators/AnimationLabel.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 628e741beae80e54a9784c9469ae59dd
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
