+++ Assets/CoreAssets/Animators/CardAnimation.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cca2212d3c20dd24e8202a80a235e39c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
