+++ Assets/CoreAssets/Animators/CardIdle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 562298b44b3084b43aa7f82cc7f6fcc2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
