+++ Assets/CoreAssets/Animators/DeckCardPrefab.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6747e1297b3b9ba4ca3174b278b95973
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
