+++ Assets/CoreAssets/Animators/TextAnimation.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 308b1f242ec35cf4caec5fc37ad7b5c6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
