+++ Assets/CoreAssets/Animators/TextIdle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 05af09430d8a2e54aae1159c17a2e32a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
