+++ Assets/CoreAssets/Prefabs/AnimationLabel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b3e2b76ac87e04046ad1948918c7f083
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
