+++ Assets/CoreAssets/Prefabs/DeckCardPrefab.prefab
@@ -222,7 +222,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d7c21276a2ab74a428ec90d628744c0e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_Text: {fileID: 2199203795218927471}
+  _label: {fileID: 2199203795218927471}
 --- !u!1 &6059513462696002114
 GameObject:
   m_ObjectHideFlags: 0
@@ -237,6 +237,7 @@ GameObject:
   - component: {fileID: 8092260227480980561}
   - component: {fileID: 4976565588218348677}
   - component: {fileID: 474686921987642790}
+  - component: {fileID: 2285170645518325015}
   m_Layer: 5
   m_Name: DeckCardPrefab
   m_TagString: Untagged
@@ -378,6 +379,28 @@ MonoBehaviour:
   _selectionMark: {fileID: 880601877099181935}
   _label: {fileID: 7972152728789023268}
   _image: {fileID: 3708233432031063951}
+  _anim: {fileID: 2285170645518325015}
+--- !u!95 &2285170645518325015
+Animator:
+  serializedVersion: 5
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6059513462696002114}
+  m_Enabled: 1
+  m_Avatar: {fileID: 0}
+  m_Controller: {fileID: 9100000, guid: 6747e1297b3b9ba4ca3174b278b95973, type: 2}
+  m_CullingMode: 0
+  m_UpdateMode: 0
+  m_ApplyRootMotion: 0
+  m_LinearVelocityBlending: 0
+  m_StabilizeFeet: 0
+  m_WarningMessage: 
+  m_HasTransformHierarchy: 1
+  m_AllowConstantClipSamplingOptimization: 1
+  m_KeepAnimatorStateOnDisable: 0
+  m_WriteDefaultValuesOnDisable: 0
 --- !u!1 &6457192169111798173
 GameObject:
   m_ObjectHideFlags: 0
@@ -600,4 +623,4 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d7c21276a2ab74a428ec90d628744c0e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_Text: {fileID: 9221249047176026155}
+  _label: {fileID: 9221249047176026155}
