+++ Assets/CoreAssets/Scripts/CoreAssetUI/View/AnimationLabelObject.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ea0a36ff0a3ad6a4eb8e4d09ff43e7ea
+guid: 2ce80d5c79372364bb0cacd10af0b313
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
