+++ Assets/CoreAssets/Scripts/CoreAssetUI/View/Component/ObservableLabelTMPro.cs
@@ -5,12 +5,14 @@ namespace CoreAssetUI.View
 {
     public class ObservableLabelTMPro : ObservableLabel
 	{
-		[SerializeField] private TMP_Text m_Text = null;
+		[SerializeField] private TMP_Text _label = null;
+
+		public TMP_Text Label => _label;
 
 		public override string Text
 		{
-			get { return m_Text.text; }
-			set { m_Text.text = value; }
+			get { return _label.text; }
+			set { _label.text = value; }
 		}
 	}
 }
