+++ Assets/CoreAssets/Scripts/CoreAssetUI/View/IBattleCardCell.cs
@@ -1,0 +1,9 @@
+using Cysharp.Threading.Tasks;
+
+namespace CoreAssetUI.View
+{
+    public interface IBattleCardCell : IDoubleTapCell
+    {
+        UniTask PlayCardAnimation();
+    }
+}
