+++ Assets/CoreAssets/Scripts/CoreAssetUI/View/IBattleCardCell.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ea0a36ff0a3ad6a4eb8e4d09ff43e7ea
+guid: a8d4c88a05141a44da70041c0a7f2436
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
