+++ Assets/CoreAssets/Scripts/CoreAssetUI/View/ISelectedCardEffectView.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace CoreAssetUI.View
+{
+    public interface ISelectedCardEffectView
+    {
+        IReadOnlyList<ICellBase> PivotList { get; }
+    }
+}
