+++ Assets/CoreAssets/Scripts/CoreAssetUI/View/ISelectedCardEffectView.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ea0a36ff0a3ad6a4eb8e4d09ff43e7ea
+guid: 667491ba0c159184d9c1ed4ccc7191d4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
