+++ Assets/CoreAssets/Scripts/CoreAssetUI/View/ISelectedCardListView.cs
@@ -1,9 +1,9 @@
-using System;
-using UnityEngine;
+using Cysharp.Threading.Tasks;
 
 namespace CoreAssetUI.Presenter
 {
     public interface ISelectedCardListView : IListView
     {
+        UniTask SetScoreEffect( int index, int val );
     }
 }
