+++ Assets/CoreAssets/Scripts/CoreAssetUI/View/ISelectedCardListView.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5e035cde4527f894bbd3c647b0242a9c
+guid: e9cb98b38317a024c8d6baa7ce2c7df8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
