+++ Assets/CoreAssets/Scripts/CoreAssetUI/View/ISelectedCardListViewOld.cs
@@ -1,0 +1,6 @@
+namespace CoreAssetUI.Presenter
+{
+    public interface ISelectedCardListViewOld : IListView
+    {
+    }
+}
