+++ Assets/CoreAssets/Scripts/CoreAssetUI/View/ISelectedCardListViewOld.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e9cb98b38317a024c8d6baa7ce2c7df8
+guid: 5e035cde4527f894bbd3c647b0242a9c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
