--- Assets/CoreAssets/Scripts/CoreAssetUI/View/IStaticCountListView.cs
@@ -1,10 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace CoreAssetUI.Presenter
-{
-    public interface IStaticCountListView : IListView
-    {
-    }
-}
