+++ Assets/CoreAssets/Scripts/CoreAssetUI/View/SelectedCardListViewOld.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ea0a36ff0a3ad6a4eb8e4d09ff43e7ea
+guid: c6f0c4b07f4b8ac41b31826112edf653
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
