+++ Assets/CoreAssets/Scripts/GameSystemSDK/BattleScene/Domain/IBattleEffectLaunchDomain.cs
@@ -10,7 +10,7 @@ namespace GameSystemSDK.BattleScene.Domain
     public interface IBattleEffectLaunchDomain
     {
         IObservable<string> OnSkillNameChanged { get; }
-        IObservable<string> OnScoreInfoChanged { get; }
+        IObservable<(int index, int score)> OnScoreInfoChanged { get; }
         IObservable<bool> OnIsEffectProccess { get; }
         UniTask RunScoreEffectProcess( IDetailScoreInfo detail, AudioClip effect );
     }
