+++ Assets/CoreAssets/Scripts/GameSystemSDK/BattleScene/Model/IBattleEffectModel.cs
@@ -10,7 +10,7 @@ namespace GameSystemSDK.BattleScene.Model
     public interface IBattleEffectModel
     {
         IObservable<string> OnSkillNameChanged { get; }
-        IObservable<string> OnScoreInfoChanged { get; }
+        IObservable<(int index, int score)> OnScoreInfoChanged { get; }
         IObservable<bool> OnIsEffectProccess { get; }
     }
 }
